Update navigation.yml
deleted Ontroduction to Linux guide which is not really a guide, just a page with Linux commands and links to videos that don't work. The list of commands as well as the link to the unix tutorials are available in the quick start guide. The onboarding video can be accessed through Training -> ARS Science tutorials.